gen3-augur/requirements.txt
biopython~=1.77
pandas~=1.0.0
setuptools~=46.0
requests~=2.22
